+++ spec/active_record_batch_consumer_spec.rb
@@ -86,7 +86,7 @@ module ActiveRecordBatchConsumerTest
           before(:each) do
             Deimos.configure do |config|
               config.schema.use_schema_classes = use_schema_classes
-              config.schema.use_full_namespace = true
+              config.avrogen.use_full_namespace = true
             end
           end