+++ spec/active_record_producer_spec.rb
@@ -95,7 +95,7 @@ describe Deimos::ActiveRecordProducer do
         before(:each) do
           Deimos.configure do |config|
             config.schema.use_schema_classes = use_schema_classes
-            config.schema.use_full_namespace = true
+            config.avrogen.use_full_namespace = true
           end
         end