+++ spec/batch_consumer_spec.rb
@@ -51,14 +51,14 @@ module ConsumerTest
 
           let(:schema_class_batch) do
             batch.map do |p|
-              Deimos::Utils::SchemaClass.instance(p, 'MySchema', 'com.my-namespace')
+              AvroGen::SchemaClass.instance(p, 'MySchema', 'com.my-namespace')
             end
           end
           let(:use_schema_classes) { true }
 
           before(:each) do
             Deimos.configure do |config|
-              config.schema.use_full_namespace = true
+              config.avrogen.use_full_namespace = true
             end
           end