+++ spec/consumer_spec.rb
@@ -48,8 +48,8 @@ module ConsumerTest
 
           before(:each) do
             Deimos.configure do |config|
-              config.schema.use_full_namespace = true
-              config.schema.schema_namespace_map = {
+              config.avrogen.use_full_namespace = true
+              config.avrogen.schema_namespace_map = {
                 'com' => 'Schemas',
                 'com.my-namespace.my-suborg' => %w(Schemas MyNamespace)
               }
@@ -62,7 +62,7 @@ module ConsumerTest
                                  'some_int' => 123 }) do |payload, _metadata|
                                    expect(payload['test_id']).to eq('foo')
                                    expect(payload['some_int']).to eq(123)
-                                 end
+            end
           end
 
           it 'should consume a nil message' do
@@ -77,7 +77,7 @@ module ConsumerTest
                                  'some_int' => 123 }) do |payload, _metadata|
                                    expect(payload['test_id']).to eq('foo')
                                    expect(payload['some_int']).to eq(123)
-                                 end
+            end
           end
 
           it 'should fail on invalid message' do
@@ -160,7 +160,7 @@ module ConsumerTest
         before(:each) do
           set_karafka_config(:use_schema_classes, true)
           Deimos.configure do |config|
-            config.schema.use_full_namespace = true
+            config.avrogen.use_full_namespace = true
           end
         end
 
@@ -196,7 +196,7 @@ module ConsumerTest
             expect(payload['test_id']).to eq('foo')
             expect(payload['some_int']).to eq(1)
             expect(payload['super_int']).to eq(9000)
-                                 end
+          end
         end
 
       end
@@ -205,8 +205,8 @@ module ConsumerTest
         before(:each) do
           set_karafka_config(:use_schema_classes, true)
           Deimos.configure do |config|
-            config.schema.use_full_namespace = true
-            config.schema.schema_namespace_map = {
+            config.avrogen.use_full_namespace = true
+            config.avrogen.schema_namespace_map = {
               'com' => 'Schemas',
               'com.my-namespace.my-suborg' => %w(Schemas MyNamespace)
             }
@@ -255,7 +255,7 @@ module ConsumerTest
             expect(payload['some_nested_record']['some_int']).to eq(1)
             expect(payload.to_h).not_to have_key('additional_field')
             expect(payload.to_h['some_nested_record']).not_to have_key('additional_field')
-                                 end
+          end
 
         end
       end